+++ lib/active_enumerable/enumerable.rb
@@ -3,7 +3,7 @@ module ActiveEnumerable
     include ::Enumerable
 
     def each(*args, &block)
-      @to_a.send(:each, *args, &block)
+      @collection.send(:each, *args, &block)
     end
   end
 end