+++ lib/active_enumerable/method_caller.rb
@@ -14,8 +14,10 @@ module ActiveEnumerable
       else
         object.public_send(method)
       end
-    rescue NoMethodError, KeyError => e
+    rescue NoMethodError => e
       raise e if raise_no_method
+    rescue KeyError => e
+      raise e, "#{e.message} for #{object}" if raise_no_method
     end
   end
 end