+++ lib/active_enumerable/scope_method.rb
@@ -1,0 +1,12 @@
+module ActiveEnumerable
+  module ScopeMethod
+    def scope(&block)
+      result = instance_exec(&block)
+      if result.is_a? Array
+        __new_relation__(result)
+      else
+        result
+      end
+    end
+  end
+end