+++ spec/queries_spec.rb
@@ -1,81 +1,27 @@
 require "spec_helper"
-require "ostruct"
 
-RSpec.describe ActiveEnumerable do
+RSpec.describe ActiveEnumerable::Queries do
 
-  class TestEnumerable
-    include ActiveEnumerable
-
-    scope :freds, -> { select { |e| e.name == "Fred" } }
-    scope :fred, -> { find_by(name: "Fred") }
+  class QueriesTestEnumerable
+    include ActiveEnumerable::Base
+    include ActiveEnumerable::Queries
   end
 
-  class OtherEnumerable
-    include ActiveEnumerable
-  end
-
-  let(:item_objects) { [OpenStruct.new(name: "Fred"), OpenStruct.new(name: "Dave")] }
-  let(:item_hashes) { [{ name: "Fred" }, { name: "Sam" }, { name: "Dave" }] }
-
-  describe ".scope" do
-
-    it "creates a scoped query collection" do
-      expect(TestEnumerable.new(item_objects).freds).to eq TestEnumerable.new([item_objects.first])
-    end
-
-    it "creates a scoped query collection" do
-      expect(TestEnumerable.new(item_objects).fred).to eq item_objects.first
-    end
-  end
-
-  describe "#where" do
-
-    it "queries collection for object with name fred" do
-      expect(TestEnumerable.new(item_objects).where(name: "Fred")).to eq TestEnumerable.new([item_objects.first])
-    end
-
-    it "queries collection for hashes with the key name and value fred" do
-      expect(TestEnumerable.new(item_hashes).where(name: "Fred")).to eq TestEnumerable.new([item_hashes.first])
-    end
-
-    it "where not" do
-      expect(TestEnumerable.new(item_hashes).where.not(name: "Fred")).to eq TestEnumerable.new([item_hashes[1],item_hashes[2]])
-    end
-
-    it "where or conditions" do
-      expect(TestEnumerable.new(item_hashes).where(name: "Dave").or(name: "Fred").where.not(name: "Sam")).
-        to eq TestEnumerable.new([item_hashes[2],item_hashes[0]])
-    end
-
-    it "where or relation" do
-      te = TestEnumerable.new(item_hashes)
-      expect(te.where(name: "Dave").or(te.where(name: "Fred")).where.not(name: "Sam")).
-        to eq TestEnumerable.new([item_hashes[2],item_hashes[0]])
-    end
-  end
+  subject { QueriesTestEnumerable.new(records) }
+  let(:records) { [{ id: 2, name: "Fred" }, { id: 4, name: "Sam" }, { id: 5, name: "Dave" }] }
 
   describe "#find_by" do
-
     it "queries collection for object with name fred" do
-      expect(TestEnumerable.new(item_objects).find_by(name: "Fred")).to eq item_objects.first
-    end
-
-    it "queries collection for hashes with the key name and value fred" do
-      expect(TestEnumerable.new(item_hashes).find_by(name: "Fred")).to eq item_hashes.first
+      expect(subject.find_by(name: "Fred")).to eq({ id: 2, name: "Fred" })
     end
   end
 
-  context "comparable" do
-    it "Objects of the same type and collection are equal" do
-      expect(TestEnumerable.new(item_objects)).to eq TestEnumerable.new(item_objects)
-    end
-
-    it "Objects of different types and same collection are not equal" do
-      expect(TestEnumerable.new(item_objects)).to_not eq OtherEnumerable.new(item_objects)
-    end
-
-    it "An ActiveEnumerable collection is equal to an array with the same item_objects" do
-      expect(TestEnumerable.new(item_objects)).to eq item_objects
-    end
+  describe "#find" do
+    it { expect(subject.find(5)).to eq({ id: 5, name: "Dave" }) }
+    it { expect(subject.find(2, 5).to_a).to eq [{ id: 2, name: "Fred" }, { id: 5, name: "Dave" }] }
+    it { expect(subject.find([2, 5]).to_a).to eq [{ id: 2, name: "Fred" }, { id: 5, name: "Dave" }] }
+    it { expect(subject.find([2]).to_a).to eq [{ id: 2, name: "Fred" }] }
+    it { expect(subject.find([2])).to be_an_instance_of QueriesTestEnumerable }
+    it { expect { subject.find(nil) }.to raise_error(ActiveEnumerable::RecordNotFound, "Couldn't find QueriesTestEnumerable without an ID") }
   end
 end