+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 0.2.0 - 2015-02-15
+### Features
+- English like querying DSL
+
+### Enhancement 
+- `#where` can match when arrays are equal.
+- `#where` can match when strings with regex.
+
 ## 0.1.0 - 2015-12-31
 This version is only to secure the rubygem name.